pyproject.toml
[project]
name = "task-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.5",
    "django>=6.0.2",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
]
